+++ naga/src/back/msl/mod.rs
@@ -235,6 +235,8 @@ pub enum Error {
     UnsupportedRayTracing,
     #[error("cooperative matrix is not supported prior to MSL 2.3")]
     UnsupportedCooperativeMatrix,
+    #[error("debugPrintf is not supported prior to MSL 3.2")]
+    UnsupportedDebugPrintf,
     #[error("overrides should not be present at this stage")]
     Override,
     #[error("bitcasting to {0:?} is not supported")]
@@ -876,6 +878,7 @@ pub fn supported_capabilities() -> crate::valid::Capabilities {
         // No MEMORY_DECORATION_VOLATILE
         | Caps::MEMORY_DECORATION_COHERENT
         | Caps::LINEAR_INTERPOLATION
+        | Caps::DEBUG_PRINTF
 }
 
 #[test]