+++ naga/src/back/spv/mod.rs
@@ -1227,4 +1227,5 @@ pub fn supported_capabilities() -> crate::valid::Capabilities {
         | Caps::MEMORY_DECORATION_COHERENT
         | Caps::MEMORY_DECORATION_VOLATILE
         | Caps::LINEAR_INTERPOLATION
+        | Caps::DEBUG_PRINTF
 }