+++ naga/src/compact/statements.rs
@@ -168,6 +168,14 @@ impl FunctionTracer<'_> {
                             self.expressions_used.insert(payload);
                         }
                     },
+                    St::DebugPrintf {
+                        format: _,
+                        ref arguments,
+                    } => {
+                        for &expr in arguments {
+                            self.expressions_used.insert(expr);
+                        }
+                    }
 
                     // Trivial statements.
                     St::Break
@@ -406,6 +414,14 @@ impl FunctionMap {
                             adjust(payload);
                         }
                     },
+                    St::DebugPrintf {
+                        format: _,
+                        ref mut arguments,
+                    } => {
+                        for expr in arguments {
+                            adjust(expr);
+                        }
+                    }
 
                     // Trivial statements.
                     St::Break