+++ naga/src/front/wgsl/error.rs
@@ -232,6 +232,7 @@ impl core::fmt::Display for ExpectedToken<'_> {
                 Token::Attribute => Kind::Str("@"),
                 Token::Number(_) => Kind::Str("number"),
                 Token::Word(s) => Kind::Str(s),
+                Token::String(_) => Kind::Str("string literal"),
                 Token::Operation(c) => Kind::FormatChar("operation (`", c, "`)"),
                 Token::LogicalOperation(c) => Kind::FormatChar("logical operation (`", c, "`)"),
                 Token::ShiftOperation(c) => {
@@ -523,6 +524,14 @@ pub(crate) enum Error<'a> {
     FunctionMustUseReturnsVoid(Span, Span),
     FunctionMustUseOnNonFunction(Span),
     InvalidWorkGroupUniformLoad(Span),
+    InvalidStringLiteral {
+        span: Span,
+        description: &'static str,
+    },
+    ExpectedStringLiteral {
+        span: Span,
+        description: &'static str,
+    },
     Internal(&'static str),
     ExpectedConstExprConcreteIntegerScalar(Span),
     ExpectedNonNegative(Span),
@@ -1347,6 +1356,16 @@ impl<'a> Error<'a> {
                 labels: vec![(*span, "".into())],
                 notes: vec!["passed type must be a workgroup pointer".into()],
                 message: "incorrect type passed to workgroupUniformLoad".into(),
+            },
+            Error::InvalidStringLiteral { span, description } => ParseError {
+                message: description.to_string().into(),
+                labels: vec![(*span, "invalid string literal".into())],
+                notes: vec![],
+            },
+            Error::ExpectedStringLiteral { span, description } => ParseError {
+                message: description.to_string().into(),
+                labels: vec![(*span, "expected a string literal".into())],
+                notes: vec![],
             },
             Error::Internal(message) => ParseError {
                 notes: vec![(*message).into()],