+++ wgpu-types/src/instance.rs
@@ -240,6 +240,15 @@ bitflags::bitflags! {
         ///
         /// When `Self::from_env()` is used takes value from `WGPU_STRICT_WEBGPU_COMPLIANCE` environment variable.
         const STRICT_WEBGPU_COMPLIANCE = 1 << 7;
+
+        /// Enable capturing the output of shader `debugPrintf`.
+        /// Supported platforms:
+        /// - Vulkan, via the `VK_LAYER_KHRONOS_validation` layer's
+        ///   ["debug printf"](https://github.com/KhronosGroup/Vulkan-ValidationLayers/blob/main/docs/debug_printf.md)
+        /// Other backends, such as Metal, capture the output without needing this flag.
+        ///
+        /// When `Self::from_env()` is used takes value from `WGPU_DEBUG_PRINTF` environment variable.
+        const DEBUG_PRINTF = 1 << 8;
     }
 }
 
@@ -298,6 +307,7 @@ impl InstanceFlags {
     /// - `WGPU_GPU_BASED_VALIDATION`
     /// - `WGPU_VALIDATION_INDIRECT_CALL`
     /// - `WGPU_STRICT_WEBGPU_COMPLIANCE`
+    /// - `WGPU_DEBUG_PRINTF`
     #[must_use]
     pub fn with_env(mut self) -> Self {
         fn env(key: &str) -> Option<bool> {
@@ -328,6 +338,9 @@ impl InstanceFlags {
         }
         if let Some(bit) = env("WGPU_STRICT_WEBGPU_COMPLIANCE") {
             self.set(Self::STRICT_WEBGPU_COMPLIANCE, bit);
+        }
+        if let Some(bit) = env("WGPU_DEBUG_PRINTF") {
+            self.set(Self::DEBUG_PRINTF, bit);
         }
 
         self